adding Login Signup switch
