+++ client/src/styles/Form.css
@@ -37,10 +37,7 @@ a{
   justify-content: center;
 }
 
-.login-form {
-    width: 23%;
-    
-}
+
 
 img {
     max-width: 300px;
